requirement.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
coverage==7.10.6
Django==5.2
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
logging==0.4.9.6
packaging==24.2
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.5.0
